+++ packages/dso-toolkit/src/components/map-message/map-message.models.ts
@@ -1,0 +1,8 @@
+export interface MapMessageActionClickEvent {}
+
+export interface MapMessage {
+  message: string;
+  variant: "success" | "error" | "instruction";
+  buttonLabels?: string[];
+  dsoActionClick?: (event: CustomEvent<MapMessageActionClickEvent>) => void;
+}